Add process utilities and prelude

+++ lib/src/time.rs
@@ -3,6 +3,6 @@
 use std::{thread, time};
 
 /// Put the program to sleep for the specified number of milliseconds.
-pub fn sleep(ms: usize) {
-    thread::sleep(time::Duration::from_millis(ms as u64));
+pub fn sleep(ms: u64) {
+    thread::sleep(time::Duration::from_millis(ms));
 }
